Modified structure: moving scheduling strategies in `utils_scheduling.py`
